@@ -1,8 +1,11 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using System;
 
 namespace PlayerTags.Data
 {
+    [Obsolete]
     [Flags]
+    [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
     public enum ActivityContext
     {
         None = 0x0,
