@@ -1,25 +1,21 @@
-﻿namespace PlayerTags.Data
+﻿using Pilz.Dalamud.ActivityContexts;
+
+namespace PlayerTags.Data
 {
     public static class ActivityContextHelper
     {
-        public static bool GetIsVisible(ActivityContext playerContext, bool desiredPveDutyVisibility, bool desiredPvpDutyVisibility, bool desiredOthersVisibility)
+        public static bool GetIsVisible(ActivityType playerContext, bool desiredPveDutyVisibility, bool desiredPvpDutyVisibility, bool desiredOthersVisibility)
         {
             bool isVisible = false;
 
-            if (playerContext.HasFlag(ActivityContext.PveDuty))
-            {
+            if (playerContext.HasFlag(ActivityType.PveDuty))
                 isVisible |= desiredPveDutyVisibility;
-            }
 
-            if (playerContext.HasFlag(ActivityContext.PvpDuty))
-            {
+            if (playerContext.HasFlag(ActivityType.PvpDuty))
                 isVisible |= desiredPvpDutyVisibility;
-            }
 
-            if (playerContext == ActivityContext.None)
-            {
+            if (playerContext == ActivityType.None)
                 isVisible |= desiredOthersVisibility;
-            }
 
             return isVisible;
         }
