@@ -1,5 +1,8 @@
-﻿namespace PlayerTags.Data
+﻿using Newtonsoft.Json;
+
+namespace PlayerTags.Data
 {
+    [JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
     public enum NameplateTitleVisibility
     {
         Default,
