@@ -1,4 +1,7 @@
-﻿namespace PlayerTags.Inheritables
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+
+namespace PlayerTags.Inheritables
 {
     public enum InheritableBehavior
     {
