@@ -1,11 +1,24 @@
 ﻿using Dalamud.Logging;
+using Dalamud.Plugin;
 using PlayerTags.Resources;
 using System;
+using System.ComponentModel;
+using System.Globalization;
 
 namespace PlayerTags
 {
     public static class Localizer
     {
+        public static void SetLanguage(string langCode)
+        {
+            SetLanguage(new CultureInfo(langCode));
+        }
+
+        public static void SetLanguage(CultureInfo cultureInfo)
+        {
+            Strings.Culture = cultureInfo;
+        }
+
         public static string GetName<TEnum>(TEnum value)
         {
             return $"{typeof(TEnum).Name}_{value}";
@@ -38,10 +51,9 @@ namespace PlayerTags
         public static string GetString(string localizedStringId)
         {
             string? value = Strings.ResourceManager.GetString(localizedStringId, Strings.Culture);
+
             if (value != null)
-            {
                 return value;
-            }
 
             PluginLog.Error($"Failed to get localized string for id {localizedStringId}");
             return localizedStringId;
