@@ -1,8 +1,15 @@
 ﻿using Dalamud.Game.Command;
+using Dalamud.Logging;
 using Dalamud.Plugin;
+using Dalamud.Plugin.Internal;
+using FFXIVClientStructs.FFXIV.Client.UI.Misc;
 using PlayerTags.Configuration;
 using PlayerTags.Data;
 using PlayerTags.Features;
+using System;
+using System.IO;
+using System.Linq;
+using System.Reflection;
 
 namespace PlayerTags
 {
@@ -15,7 +22,6 @@ namespace PlayerTags
         private PluginData m_PluginData;
         private PluginConfigurationUI m_PluginConfigurationUI;
 
-        private LinkSelfInChatFeature m_LinkSelfInChatFeature;
         private CustomTagsContextMenuFeature m_CustomTagsContextMenuFeature;
         private NameplateTagTargetFeature m_NameplatesTagTargetFeature;
         private ChatTagTargetFeature m_ChatTagTargetFeature;
@@ -23,18 +29,21 @@ namespace PlayerTags
         public Plugin(DalamudPluginInterface pluginInterface)
         {
             PluginServices.Initialize(pluginInterface);
+            Pilz.Dalamud.PluginServices.Initialize(pluginInterface);
 
-            m_PluginConfiguration = PluginServices.DalamudPluginInterface.GetPluginConfig() as PluginConfiguration ?? new PluginConfiguration();
+            m_PluginConfiguration = PluginConfiguration.LoadPluginConfig() ?? new PluginConfiguration();
             m_PluginData = new PluginData(m_PluginConfiguration);
             m_PluginConfigurationUI = new PluginConfigurationUI(m_PluginConfiguration, m_PluginData);
 
+            Localizer.SetLanguage(PluginServices.DalamudPluginInterface.UiLanguage);
+            PluginServices.DalamudPluginInterface.LanguageChanged += DalamudPluginInterface_LanguageChanged;
+
             PluginServices.DalamudPluginInterface.UiBuilder.Draw += UiBuilder_Draw;
             PluginServices.DalamudPluginInterface.UiBuilder.OpenConfigUi += UiBuilder_OpenConfigUi;
-            PluginServices.CommandManager.AddHandler(c_CommandName, new CommandInfo((string command, string arguments) =>
+            PluginServices.CommandManager.AddHandler(c_CommandName, new CommandInfo((string command, string arguments) => UiBuilder_OpenConfigUi())
             {
-                UiBuilder_OpenConfigUi();
-            }) { HelpMessage = "Shows the config" });
-            m_LinkSelfInChatFeature = new LinkSelfInChatFeature(m_PluginConfiguration, m_PluginData);
+                HelpMessage = Resources.Strings.Loc_Command_playertags
+            });
             m_CustomTagsContextMenuFeature = new CustomTagsContextMenuFeature(m_PluginConfiguration, m_PluginData);
             m_NameplatesTagTargetFeature = new NameplateTagTargetFeature(m_PluginConfiguration, m_PluginData);
             m_ChatTagTargetFeature = new ChatTagTargetFeature(m_PluginConfiguration, m_PluginData);
@@ -45,18 +54,21 @@ namespace PlayerTags
             m_ChatTagTargetFeature.Dispose();
             m_NameplatesTagTargetFeature.Dispose();
             m_CustomTagsContextMenuFeature.Dispose();
-            m_LinkSelfInChatFeature.Dispose();
+            PluginServices.DalamudPluginInterface.LanguageChanged -= DalamudPluginInterface_LanguageChanged;
             PluginServices.CommandManager.RemoveHandler(c_CommandName);
             PluginServices.DalamudPluginInterface.UiBuilder.OpenConfigUi -= UiBuilder_OpenConfigUi;
             PluginServices.DalamudPluginInterface.UiBuilder.Draw -= UiBuilder_Draw;
         }
 
+        private void DalamudPluginInterface_LanguageChanged(string langCode)
+        {
+            Localizer.SetLanguage(langCode);
+        }
+
         private void UiBuilder_Draw()
         {
             if (m_PluginConfiguration.IsVisible)
-            {
                 m_PluginConfigurationUI.Draw();
-            }
         }
 
         private void UiBuilder_OpenConfigUi()
