@@ -1,10 +1,10 @@
 ﻿//------------------------------------------------------------------------------
 // <auto-generated>
-//     This code was generated by a tool.
-//     Runtime Version:4.0.30319.42000
+//     Dieser Code wurde von einem Tool generiert.
+//     Laufzeitversion:4.0.30319.42000
 //
-//     Changes to this file may cause incorrect behavior and will be lost if
-//     the code is regenerated.
+//     Änderungen an dieser Datei können falsches Verhalten verursachen und gehen verloren, wenn
+//     der Code erneut generiert wird.
 // </auto-generated>
 //------------------------------------------------------------------------------
 
@@ -13,12 +13,12 @@ namespace PlayerTags.Resources {
     
     
     /// <summary>
-    ///   A strongly-typed resource class, for looking up localized strings, etc.
+    ///   Eine stark typisierte Ressourcenklasse zum Suchen von lokalisierten Zeichenfolgen usw.
     /// </summary>
-    // This class was auto-generated by the StronglyTypedResourceBuilder
-    // class via a tool like ResGen or Visual Studio.
-    // To add or remove a member, edit your .ResX file then rerun ResGen
-    // with the /str option, or rebuild your VS project.
+    // Diese Klasse wurde von der StronglyTypedResourceBuilder automatisch generiert
+    // -Klasse über ein Tool wie ResGen oder Visual Studio automatisch generiert.
+    // Um einen Member hinzuzufügen oder zu entfernen, bearbeiten Sie die .ResX-Datei und führen dann ResGen
+    // mit der /str-Option erneut aus, oder Sie erstellen Ihr VS-Projekt neu.
     [global::System.CodeDom.Compiler.GeneratedCodeAttribute("System.Resources.Tools.StronglyTypedResourceBuilder", "17.0.0.0")]
     [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
     [global::System.Runtime.CompilerServices.CompilerGeneratedAttribute()]
@@ -33,7 +33,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Returns the cached ResourceManager instance used by this class.
+        ///   Gibt die zwischengespeicherte ResourceManager-Instanz zurück, die von dieser Klasse verwendet wird.
         /// </summary>
         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Advanced)]
         public static global::System.Resources.ResourceManager ResourceManager {
@@ -47,8 +47,8 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Overrides the current thread's CurrentUICulture property for all
-        ///   resource lookups using this strongly typed resource class.
+        ///   Überschreibt die CurrentUICulture-Eigenschaft des aktuellen Threads für alle
+        ///   Ressourcenzuordnungen, die diese stark typisierte Ressourcenklasse verwenden.
         /// </summary>
         [global::System.ComponentModel.EditorBrowsableAttribute(global::System.ComponentModel.EditorBrowsableState.Advanced)]
         public static global::System.Globalization.CultureInfo Culture {
@@ -61,7 +61,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Activity properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Activity properties ähnelt.
         /// </summary>
         public static string Loc_ActivityCategory {
             get {
@@ -70,7 +70,88 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Custom.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Below options to be apply for ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Everywhere ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_All {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_All", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die The most options below will apply everywhere you are. They will be the same in Overworld, in PvE Duties and PvP Duties. Your options defined specifically for another context will be overwritten. ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_All_Description {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_All_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die No Duty ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_None {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_None", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die The most options below will apply only if you are outside of any duty. ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_None_Description {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_None_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die PvE Duties ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_PveDuty {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_PveDuty", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die The most options below will apply only to PvE Duties. ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_PveDuty_Description {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_PveDuty_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die PvP Duties ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_PvpDuty {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_PvpDuty", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die The most options below will apply only to PvP Duties. ähnelt.
+        /// </summary>
+        public static string Loc_ActivityContextSelection_PvpDuty_Description {
+            get {
+                return ResourceManager.GetString("Loc_ActivityContextSelection_PvpDuty_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Custom ähnelt.
         /// </summary>
         public static string Loc_AllCustomTags {
             get {
@@ -79,7 +160,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Roles.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Roles ähnelt.
         /// </summary>
         public static string Loc_AllRoleTags {
             get {
@@ -88,7 +169,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to All.
+        ///   Sucht eine lokalisierte Zeichenfolge, die All ähnelt.
         /// </summary>
         public static string Loc_AllTags {
             get {
@@ -97,7 +178,25 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Custom id.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Advanced Chat Options ähnelt.
+        /// </summary>
+        public static string Loc_ChatFeatureCategory {
+            get {
+                return ResourceManager.GetString("Loc_ChatFeatureCategory", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Shows the config window for Player Tags ähnelt.
+        /// </summary>
+        public static string Loc_Command_playertags {
+            get {
+                return ResourceManager.GetString("Loc_Command_playertags", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Custom id ähnelt.
         /// </summary>
         public static string Loc_CustomId {
             get {
@@ -106,7 +205,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Custom.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Custom ähnelt.
         /// </summary>
         public static string Loc_CustomTags {
             get {
@@ -115,7 +214,168 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Melee.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Handling for dead players ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Gray out ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling_GrayOut {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling_GrayOut", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply any tag but gray out the nameplate. ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling_GrayOut_Description {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling_GrayOut_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Ignore ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling_Ignore {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling_Ignore", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Don&apos;t process dead players and don&apos;t apply any tag for. ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling_Ignore_Description {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling_Ignore_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Include ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling_Include {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling_Include", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Handle dead players as they are alive. No difference between dead and alive players. ähnelt.
+        /// </summary>
+        public static string Loc_DeadPlayerHandling_Include_Description {
+            get {
+                return ResourceManager.GetString("Loc_DeadPlayerHandling_Include_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Template ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Basic ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Basic {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Basic", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Same as Empty, but includes a very basic pre-configuration for formatting and coloring.
+        ///Can also be used if you want to make your own configuration but want basic formatting. ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Basic_Description {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Basic_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Here you can choose the template for all the settings below.
+        ///The template is used as basic set of properties. Every change you are making is what get saved. But all options from this template will not be saved.
+        ///This helps you by not needing to overwrite so much properties - or by doing your completely own configuration without a template.
+        ///
+        ///Warning:
+        ///Changing this can cause properties to be resetted. You may loose a part of your configuration.
+        ///After you changed this, ensure everything is setted up how y [Rest der Zeichenfolge wurde abgeschnitten]&quot;; ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Description {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Full ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Full {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Full", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show the job tag with color and also color the player name element. ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Full_Description {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Full_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Empty ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_None {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_None", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die No single configuration made. This is a completely empty template.
+        ///Use this if you want to to have every option under your control or just want to make only a few configurations. ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_None_Description {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_None_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Simple ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Simple {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Simple", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Shows the job tag with color and the role icon by replacing the title. ähnelt.
+        /// </summary>
+        public static string Loc_DefaultPluginDataTemplate_Simple_Description {
+            get {
+                return ResourceManager.GetString("Loc_DefaultPluginDataTemplate_Simple_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Melee ähnelt.
         /// </summary>
         public static string Loc_DpsRole_Melee {
             get {
@@ -124,7 +384,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Ranged.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Ranged ähnelt.
         /// </summary>
         public static string Loc_DpsRole_Ranged {
             get {
@@ -133,7 +393,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Add to players.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Add to players ähnelt.
         /// </summary>
         public static string Loc_GameObjectNamesToApplyTo {
             get {
@@ -142,7 +402,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to A list of players to add tags to, separated by commas or semi-colons. E.g. &quot;Cloud Strife, Tifa Lockhart&quot;..
+        ///   Sucht eine lokalisierte Zeichenfolge, die A list of players to add tags to, separated by commas or semi-colons. E.g. &quot;Cloud Strife, Tifa Lockhart&quot;. ähnelt.
         /// </summary>
         public static string Loc_GameObjectNamesToApplyTo_Description {
             get {
@@ -151,7 +411,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to General properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die General properties ähnelt.
         /// </summary>
         public static string Loc_GeneralCategory {
             get {
@@ -160,7 +420,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Icon.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Icon ähnelt.
         /// </summary>
         public static string Loc_Icon {
             get {
@@ -169,7 +429,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The icon that will be displayed..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The icon that will be displayed. ähnelt.
         /// </summary>
         public static string Loc_Icon_Description {
             get {
@@ -178,7 +438,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Icon properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Icon properties ähnelt.
         /// </summary>
         public static string Loc_IconCategory {
             get {
@@ -187,7 +447,25 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Apply tags to all chat messages.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Insert behind group number prefix in chat ähnelt.
+        /// </summary>
+        public static string Loc_InsertBehindNumberPrefixInChat {
+            get {
+                return ResourceManager.GetString("Loc_InsertBehindNumberPrefixInChat", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die If enabled, the Tag and Icon will get inserted behind the Group/Alliance number prefix in Chat instead of before.  ähnelt.
+        /// </summary>
+        public static string Loc_InsertBehindNumberPrefixInChat_Description {
+            get {
+                return ResourceManager.GetString("Loc_InsertBehindNumberPrefixInChat_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply tags to all chat messages ähnelt.
         /// </summary>
         public static string Loc_IsApplyTagsToAllChatMessagesEnabled {
             get {
@@ -196,7 +474,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Applies tags to all chat messages, including non-social messages..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Applies tags to all chat messages, including non-social messages. ähnelt.
         /// </summary>
         public static string Loc_IsApplyTagsToAllChatMessagesEnabled_Description {
             get {
@@ -205,7 +483,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Context menu integration.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Context menu integration ähnelt.
         /// </summary>
         public static string Loc_IsCustomTagsContextMenuEnabled {
             get {
@@ -214,7 +492,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Options will be available in context menus for adding and removing custom tags from players..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Options will be available in context menus for adding and removing custom tags from players. ähnelt.
         /// </summary>
         public static string Loc_IsCustomTagsContextMenuEnabled_Description {
             get {
@@ -223,7 +501,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Enabled.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Enabled ähnelt.
         /// </summary>
         public static string Loc_IsEnabled {
             get {
@@ -232,7 +510,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Enable the value of this override..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Enable the value of this override. ähnelt.
         /// </summary>
         public static string Loc_IsEnabled_Description {
             get {
@@ -241,7 +519,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Expanded.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Expanded ähnelt.
         /// </summary>
         public static string Loc_IsExpanded {
             get {
@@ -250,7 +528,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show in chat.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show in chat ähnelt.
         /// </summary>
         public static string Loc_IsIconVisibleInChat {
             get {
@@ -259,34 +537,25 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the icon will be shown in chat..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show job icon in nameplates ähnelt.
         /// </summary>
-        public static string Loc_IsIconVisibleInChat_Description {
+        public static string Loc_IsJobIconVisibleInNameplates {
             get {
-                return ResourceManager.GetString("Loc_IsIconVisibleInChat_Description", resourceCulture);
+                return ResourceManager.GetString("Loc_IsJobIconVisibleInNameplates", resourceCulture);
             }
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show in nameplates.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the job icon will be shown in nameplates. ähnelt.
         /// </summary>
-        public static string Loc_IsIconVisibleInNameplates {
+        public static string Loc_IsJobIconVisibleInNameplates_Description {
             get {
-                return ResourceManager.GetString("Loc_IsIconVisibleInNameplates", resourceCulture);
+                return ResourceManager.GetString("Loc_IsJobIconVisibleInNameplates_Description", resourceCulture);
             }
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the icon will be shown in nameplates..
-        /// </summary>
-        public static string Loc_IsIconVisibleInNameplates_Description {
-            get {
-                return ResourceManager.GetString("Loc_IsIconVisibleInNameplates_Description", resourceCulture);
-            }
-        }
-        
-        /// <summary>
-        ///   Looks up a localized string similar to Apply tags to self in chat.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply tags to self in chat ähnelt.
         /// </summary>
         public static string Loc_IsLinkSelfInChatEnabled {
             get {
@@ -295,7 +564,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Attempts to link your chat name to your character, allowing tags to be applied to self in chat..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Attempts to link your chat name to your character, allowing tags to be applied to self in chat. ähnelt.
         /// </summary>
         public static string Loc_IsLinkSelfInChatEnabled_Description {
             get {
@@ -304,7 +573,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Randomly generate player names.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Randomly generate player names ähnelt.
         /// </summary>
         public static string Loc_IsPlayerNameRandomlyGenerated {
             get {
@@ -313,7 +582,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Replace every player&apos;s name with a randomly generated one..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Replace every player&apos;s name with a randomly generated one. ähnelt.
         /// </summary>
         public static string Loc_IsPlayerNameRandomlyGenerated_Description {
             get {
@@ -322,7 +591,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show alliance members.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show alliance members ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabAllianceVisible {
             get {
@@ -331,7 +600,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show alliance members in the players list..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show alliance members in the players list. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabAllianceVisible_Description {
             get {
@@ -340,7 +609,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show enemies.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show enemies ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabEnemiesVisible {
             get {
@@ -349,7 +618,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show enemies in the players list..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show enemies in the players list. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabEnemiesVisible_Description {
             get {
@@ -358,7 +627,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show friends.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show friends ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabFriendsVisible {
             get {
@@ -367,7 +636,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show friends in the players list..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show friends in the players list. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabFriendsVisible_Description {
             get {
@@ -376,7 +645,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Order by proximity.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Order by proximity ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabOrderedByProximity {
             get {
@@ -385,7 +654,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Players that are closer to you will be ordered towards the top..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Players that are closer to you will be ordered towards the top. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabOrderedByProximity_Description {
             get {
@@ -394,7 +663,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show others.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show others ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabOthersVisible {
             get {
@@ -403,7 +672,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show others in the players list..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show others in the players list. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabOthersVisible_Description {
             get {
@@ -412,7 +681,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show party members.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show party members ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabPartyVisible {
             get {
@@ -421,7 +690,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show party members in the players list..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show party members in the players list. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabPartyVisible_Description {
             get {
@@ -430,7 +699,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show self.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show self ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabSelfVisible {
             get {
@@ -439,7 +708,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show yourself in the players list..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show yourself in the players list. ähnelt.
         /// </summary>
         public static string Loc_IsPlayersTabSelfVisible_Description {
             get {
@@ -448,7 +717,52 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Selected.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show role icon in Chat ähnelt.
+        /// </summary>
+        public static string Loc_IsRoleIconVisibleInChat {
+            get {
+                return ResourceManager.GetString("Loc_IsRoleIconVisibleInChat", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the role icon will be shown in chat. ähnelt.
+        /// </summary>
+        public static string Loc_IsRoleIconVisibleInChat_Description {
+            get {
+                return ResourceManager.GetString("Loc_IsRoleIconVisibleInChat_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show role icon in nameplates ähnelt.
+        /// </summary>
+        public static string Loc_IsRoleIconVisibleInNameplates {
+            get {
+                return ResourceManager.GetString("Loc_IsRoleIconVisibleInNameplates", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the role icon will be shown in nameplates. ähnelt.
+        /// </summary>
+        public static string Loc_IsRoleIconVisibleInNameplates_Description {
+            get {
+                return ResourceManager.GetString("Loc_IsRoleIconVisibleInNameplates_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show role icon in nameplate ähnelt.
+        /// </summary>
+        public static string Loc_IsRoleJobIconVisibleInNameplates {
+            get {
+                return ResourceManager.GetString("Loc_IsRoleJobIconVisibleInNameplates", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Selected ähnelt.
         /// </summary>
         public static string Loc_IsSelected {
             get {
@@ -457,7 +771,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show inherited properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show inherited properties ähnelt.
         /// </summary>
         public static string Loc_IsShowInheritedPropertiesEnabled {
             get {
@@ -466,7 +780,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Properties inherited from parent nodes will be shown in the editor..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Properties inherited from parent nodes will be shown in the editor. ähnelt.
         /// </summary>
         public static string Loc_IsShowInheritedPropertiesEnabled_Description {
             get {
@@ -475,7 +789,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Apply color to chat name.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply color to chat name ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToChatName {
             get {
@@ -484,7 +798,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the color will be applied to the name in chat..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the color will be applied to the name in chat. ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToChatName_Description {
             get {
@@ -493,7 +807,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Apply color to nameplate free company.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply color to nameplate free company ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToNameplateFreeCompany {
             get {
@@ -502,7 +816,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the color will be applied to the free company in nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the color will be applied to the free company in nameplates. ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToNameplateFreeCompany_Description {
             get {
@@ -511,7 +825,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Apply color to nameplate name.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply color to nameplate name ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToNameplateName {
             get {
@@ -520,7 +834,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the color will be applied to the name in nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the color will be applied to the name in nameplates. ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToNameplateName_Description {
             get {
@@ -529,7 +843,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Apply color to nameplate title.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Apply color to nameplate title ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToNameplateTitle {
             get {
@@ -538,7 +852,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the color will be applied to title in nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the color will be applied to title in nameplates. ähnelt.
         /// </summary>
         public static string Loc_IsTextColorAppliedToNameplateTitle_Description {
             get {
@@ -547,7 +861,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Italic.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Italic ähnelt.
         /// </summary>
         public static string Loc_IsTextItalic {
             get {
@@ -556,7 +870,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the text will be italic..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the text will be italic. ähnelt.
         /// </summary>
         public static string Loc_IsTextItalic_Description {
             get {
@@ -565,7 +879,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show in chat.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show in chat ähnelt.
         /// </summary>
         public static string Loc_IsTextVisibleInChat {
             get {
@@ -574,7 +888,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the text will be shown in chat..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the text will be shown in chat. ähnelt.
         /// </summary>
         public static string Loc_IsTextVisibleInChat_Description {
             get {
@@ -583,7 +897,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show in nameplates.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show in nameplates ähnelt.
         /// </summary>
         public static string Loc_IsTextVisibleInNameplates {
             get {
@@ -592,7 +906,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the text will be shown in nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the text will be shown in nameplates. ähnelt.
         /// </summary>
         public static string Loc_IsTextVisibleInNameplates_Description {
             get {
@@ -601,7 +915,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show for alliance members.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show for alliance members ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForAlliancePlayers {
             get {
@@ -610,7 +924,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible for alliance members who are not in the current party..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible for alliance members who are not in the current party. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForAlliancePlayers_Description {
             get {
@@ -619,7 +933,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show for enemies.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show for enemies ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForEnemyPlayers {
             get {
@@ -628,7 +942,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible for enemies in pvp..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible for enemies in pvp. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForEnemyPlayers_Description {
             get {
@@ -637,7 +951,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show for friends.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show for friends ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForFriendPlayers {
             get {
@@ -646,7 +960,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible for friends..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible for friends. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForFriendPlayers_Description {
             get {
@@ -655,7 +969,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show for others.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show for others ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForOtherPlayers {
             get {
@@ -664,7 +978,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible for players in other circumstances for which there is no specific option..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible for players in other circumstances for which there is no specific option. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForOtherPlayers_Description {
             get {
@@ -673,7 +987,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show for party members.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show for party members ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForPartyPlayers {
             get {
@@ -682,7 +996,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible for party members..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible for party members. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForPartyPlayers_Description {
             get {
@@ -691,7 +1005,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show for self.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show for self ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForSelf {
             get {
@@ -700,7 +1014,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible for the local player..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible for the local player. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleForSelf_Description {
             get {
@@ -709,7 +1023,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show elsewhere.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show elsewhere ähnelt.
         /// </summary>
         public static string Loc_IsVisibleInOverworld {
             get {
@@ -718,7 +1032,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible in other circumstances for which there is no specific option..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible in other circumstances for which there is no specific option. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleInOverworld_Description {
             get {
@@ -727,7 +1041,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show in pve duties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show in pve duties ähnelt.
         /// </summary>
         public static string Loc_IsVisibleInPveDuties {
             get {
@@ -736,7 +1050,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible in pve duties..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible in pve duties. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleInPveDuties_Description {
             get {
@@ -745,7 +1059,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Show in pvp duties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Show in pvp duties ähnelt.
         /// </summary>
         public static string Loc_IsVisibleInPvpDuties {
             get {
@@ -754,7 +1068,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Whether the tag should be visible in pvp duties..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Whether the tag should be visible in pvp duties. ähnelt.
         /// </summary>
         public static string Loc_IsVisibleInPvpDuties_Description {
             get {
@@ -763,7 +1077,25 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Hand.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Job icon set ähnelt.
+        /// </summary>
+        public static string Loc_JobIconSet {
+            get {
+                return ResourceManager.GetString("Loc_JobIconSet", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die The icon set to use for displaying the job icon. You can also choose the role icon set to display the role icon instead. ähnelt.
+        /// </summary>
+        public static string Loc_JobIconSet_Description {
+            get {
+                return ResourceManager.GetString("Loc_JobIconSet_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Hand ähnelt.
         /// </summary>
         public static string Loc_LandHandRole_Hand {
             get {
@@ -772,7 +1104,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Land.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Land ähnelt.
         /// </summary>
         public static string Loc_LandHandRole_Land {
             get {
@@ -781,7 +1113,34 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Free company.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Move Status Icon to Nameplate text if possible ähnelt.
+        /// </summary>
+        public static string Loc_MoveStatusIconToNameplateTextIfPossible {
+            get {
+                return ResourceManager.GetString("Loc_MoveStatusIconToNameplateTextIfPossible", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die If the current status icon is available as Font Icon then move it to the Player Name text in the Nameplate, so there is place for another icon to use, e.g. the job icon. ähnelt.
+        /// </summary>
+        public static string Loc_MoveStatusIconToNameplateTextIfPossible_Description {
+            get {
+                return ResourceManager.GetString("Loc_MoveStatusIconToNameplateTextIfPossible_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Nameplate properties ähnelt.
+        /// </summary>
+        public static string Loc_NameplateCategory {
+            get {
+                return ResourceManager.GetString("Loc_NameplateCategory", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Free company ähnelt.
         /// </summary>
         public static string Loc_NameplateElement_FreeCompany {
             get {
@@ -790,7 +1149,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The free company element on nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The free company element on nameplates. ähnelt.
         /// </summary>
         public static string Loc_NameplateElement_FreeCompany_Description {
             get {
@@ -799,7 +1158,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Name.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Name ähnelt.
         /// </summary>
         public static string Loc_NameplateElement_Name {
             get {
@@ -808,7 +1167,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The name element on nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The name element on nameplates. ähnelt.
         /// </summary>
         public static string Loc_NameplateElement_Name_Description {
             get {
@@ -817,7 +1176,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Title.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Title ähnelt.
         /// </summary>
         public static string Loc_NameplateElement_Title {
             get {
@@ -826,7 +1185,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title element on nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title element on nameplates. ähnelt.
         /// </summary>
         public static string Loc_NameplateElement_Title_Description {
             get {
@@ -835,7 +1194,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Free company visibility.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Free company visibility ähnelt.
         /// </summary>
         public static string Loc_NameplateFreeCompanyVisibility {
             get {
@@ -844,7 +1203,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Default.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Default ähnelt.
         /// </summary>
         public static string Loc_NameplateFreeCompanyVisibility_Default {
             get {
@@ -853,7 +1212,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The free company on nameplates will only be visible when the character is in a free company..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The free company on nameplates will only be visible when the character is in a free company. ähnelt.
         /// </summary>
         public static string Loc_NameplateFreeCompanyVisibility_Default_Description {
             get {
@@ -862,7 +1221,16 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Never visible.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Defines whenever the free company element of the nameplate should be visible or not. ähnelt.
+        /// </summary>
+        public static string Loc_NameplateFreeCompanyVisibility_Description {
+            get {
+                return ResourceManager.GetString("Loc_NameplateFreeCompanyVisibility_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Never visible ähnelt.
         /// </summary>
         public static string Loc_NameplateFreeCompanyVisibility_Never {
             get {
@@ -871,7 +1239,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The free company on nameplates will never be visible..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The free company on nameplates will never be visible. ähnelt.
         /// </summary>
         public static string Loc_NameplateFreeCompanyVisibility_Never_Description {
             get {
@@ -880,7 +1248,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Title position.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Title position ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition {
             get {
@@ -889,7 +1257,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Always above name.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Always above name ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition_AlwaysAboveName {
             get {
@@ -898,7 +1266,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will always be positioned above the name..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will always be positioned above the name. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition_AlwaysAboveName_Description {
             get {
@@ -907,7 +1275,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Always below name.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Always below name ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition_AlwaysBelowName {
             get {
@@ -916,7 +1284,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will always be positioned below the name..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will always be positioned below the name. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition_AlwaysBelowName_Description {
             get {
@@ -925,7 +1293,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Default.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Default ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition_Default {
             get {
@@ -934,7 +1302,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will be positioned depending on the title..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will be positioned depending on the title. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitlePosition_Default_Description {
             get {
@@ -943,7 +1311,16 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Title visibility.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Defines the position for the title element of the nameplate. Should it be above or below the name? You can decide! ähnelt.
+        /// </summary>
+        public static string Loc_NameplateTitlePosition_Description {
+            get {
+                return ResourceManager.GetString("Loc_NameplateTitlePosition_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Title visibility ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility {
             get {
@@ -952,7 +1329,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Always visible.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Always visible ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_Always {
             get {
@@ -961,7 +1338,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will always be visible, even when the character does not have a title..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will always be visible, even when the character does not have a title. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_Always_Description {
             get {
@@ -970,7 +1347,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Default.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Default ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_Default {
             get {
@@ -979,7 +1356,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will only be visible when the character has a title..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will only be visible when the character has a title. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_Default_Description {
             get {
@@ -988,7 +1365,16 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Never visible.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Defines whenever the title element of the nameplate should be visible or not. ähnelt.
+        /// </summary>
+        public static string Loc_NameplateTitleVisibility_Description {
+            get {
+                return ResourceManager.GetString("Loc_NameplateTitleVisibility_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Never visible ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_Never {
             get {
@@ -997,7 +1383,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will never visible..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will never visible. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_Never_Description {
             get {
@@ -1006,7 +1392,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Only visible when it has tags.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Only visible when it has tags ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_WhenHasTags {
             get {
@@ -1015,7 +1401,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The title on nameplates will only be visible when it has tags..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The title on nameplates will only be visible when it has tags. ähnelt.
         /// </summary>
         public static string Loc_NameplateTitleVisibility_WhenHasTags_Description {
             get {
@@ -1024,7 +1410,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Player properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Player properties ähnelt.
         /// </summary>
         public static string Loc_PlayerCategory {
             get {
@@ -1033,7 +1419,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Position properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Position properties ähnelt.
         /// </summary>
         public static string Loc_PositionCategory {
             get {
@@ -1042,7 +1428,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Magical.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Magical ähnelt.
         /// </summary>
         public static string Loc_RangedDpsRole_Magical {
             get {
@@ -1051,7 +1437,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Physical.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Physical ähnelt.
         /// </summary>
         public static string Loc_RangedDpsRole_Physical {
             get {
@@ -1060,7 +1446,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to DPS.
+        ///   Sucht eine lokalisierte Zeichenfolge, die DPS ähnelt.
         /// </summary>
         public static string Loc_Role_Dps {
             get {
@@ -1069,7 +1455,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Healer.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Healer ähnelt.
         /// </summary>
         public static string Loc_Role_Healer {
             get {
@@ -1078,7 +1464,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Land/Hand.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Land/Hand ähnelt.
         /// </summary>
         public static string Loc_Role_LandHand {
             get {
@@ -1087,7 +1473,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Tank.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Tank ähnelt.
         /// </summary>
         public static string Loc_Role_Tank {
             get {
@@ -1096,7 +1482,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Add a custom tag..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Add a custom tag. ähnelt.
         /// </summary>
         public static string Loc_Static_AddCustomTag_Description {
             get {
@@ -1105,7 +1491,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Add a property override..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Add a property override. ähnelt.
         /// </summary>
         public static string Loc_Static_AddPropertyOverride_Description {
             get {
@@ -1114,7 +1500,25 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Add tag.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Chat ähnelt.
+        /// </summary>
+        public static string Loc_Static_Chat {
+            get {
+                return ResourceManager.GetString("Loc_Static_Chat", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Chat (Experimental) ähnelt.
+        /// </summary>
+        public static string Loc_Static_ChatExperimental {
+            get {
+                return ResourceManager.GetString("Loc_Static_ChatExperimental", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Add tag: {0} ähnelt.
         /// </summary>
         public static string Loc_Static_ContextMenu_AddTag {
             get {
@@ -1123,7 +1527,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Remove tag.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Remove tag: {0} ähnelt.
         /// </summary>
         public static string Loc_Static_ContextMenu_RemoveTag {
             get {
@@ -1132,7 +1536,16 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Development.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Current Activity Profile ähnelt.
+        /// </summary>
+        public static string Loc_Static_CurrentActivityProfile {
+            get {
+                return ResourceManager.GetString("Loc_Static_CurrentActivityProfile", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Development ähnelt.
         /// </summary>
         public static string Loc_Static_Development {
             get {
@@ -1141,7 +1554,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Experimental.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Experimental ähnelt.
         /// </summary>
         public static string Loc_Static_Experimental {
             get {
@@ -1150,7 +1563,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Add &apos;{0}&apos; to {1}..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Add &apos;{0}&apos; to {1}. ähnelt.
         /// </summary>
         public static string Loc_Static_Format_AddTagToPlayer {
             get {
@@ -1159,7 +1572,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to General.
+        ///   Sucht eine lokalisierte Zeichenfolge, die General ähnelt.
         /// </summary>
         public static string Loc_Static_General {
             get {
@@ -1168,7 +1581,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to &lt;Inherited&gt;.
+        ///   Sucht eine lokalisierte Zeichenfolge, die &lt;Inherited&gt; ähnelt.
         /// </summary>
         public static string Loc_Static_Inherited {
             get {
@@ -1177,7 +1590,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Nameplates.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Nameplates ähnelt.
         /// </summary>
         public static string Loc_Static_Nameplates {
             get {
@@ -1186,7 +1599,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to New tag.
+        ///   Sucht eine lokalisierte Zeichenfolge, die New tag ähnelt.
         /// </summary>
         public static string Loc_Static_NewTag {
             get {
@@ -1195,7 +1608,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to &lt;No text&gt;.
+        ///   Sucht eine lokalisierte Zeichenfolge, die &lt;No text&gt; ähnelt.
         /// </summary>
         public static string Loc_Static_NoText {
             get {
@@ -1204,7 +1617,16 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Player.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Other (Experimental) ähnelt.
+        /// </summary>
+        public static string Loc_Static_OtherExperimental {
+            get {
+                return ResourceManager.GetString("Loc_Static_OtherExperimental", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Player ähnelt.
         /// </summary>
         public static string Loc_Static_PlayerName {
             get {
@@ -1213,7 +1635,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Player Tags.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Player Tags ähnelt.
         /// </summary>
         public static string Loc_Static_PluginName {
             get {
@@ -1222,7 +1644,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Quick Tag.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Quick Tag ähnelt.
         /// </summary>
         public static string Loc_Static_QuickTag {
             get {
@@ -1231,7 +1653,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Remove this custom tag..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Remove this custom tag. ähnelt.
         /// </summary>
         public static string Loc_Static_RemoveCustomTag_Description {
             get {
@@ -1240,7 +1662,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Remove this property override. The value will be inherited from the parent item..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Remove this property override. The value will be inherited from the parent item. ähnelt.
         /// </summary>
         public static string Loc_Static_RemovePropertyOverride_Description {
             get {
@@ -1249,7 +1671,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Reset all items to their defaults. Custom tags will not be touched..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Reset all items to their defaults. Custom tags will not be touched. ähnelt.
         /// </summary>
         public static string Loc_Static_ResetAllDefault_Description {
             get {
@@ -1258,7 +1680,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Reset this item to its defaults..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Reset this item to its defaults. ähnelt.
         /// </summary>
         public static string Loc_Static_ResetDefault_Description {
             get {
@@ -1267,7 +1689,16 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Tagged Players.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Status Icon Priorizer ähnelt.
+        /// </summary>
+        public static string Loc_Static_StatusIconPrioList {
+            get {
+                return ResourceManager.GetString("Loc_Static_StatusIconPrioList", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Tagged Players ähnelt.
         /// </summary>
         public static string Loc_Static_TaggedPlayers {
             get {
@@ -1276,7 +1707,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Tags.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Tags ähnelt.
         /// </summary>
         public static string Loc_Static_Tags {
             get {
@@ -1285,7 +1716,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to This plugin can modify nameplates and chat. When combined with other plugins that modify these things, you are likely to see unexpected behavior. The load order of your plugins may make a difference..
+        ///   Sucht eine lokalisierte Zeichenfolge, die This plugin can modify nameplates and chat. When combined with other plugins that modify these things, you are likely to see unexpected behavior. The load order of your plugins may make a difference. ähnelt.
         /// </summary>
         public static string Loc_Static_WarningMessage {
             get {
@@ -1294,7 +1725,241 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to After.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Reset to default ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizer_ResetToDefault {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizer_ResetToDefault", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Resets all condition sets to the default settings ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizer_ResetToDefault_Description {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizer_ResetToDefault_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Reset to empty ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizer_ResetToEmpty {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizer_ResetToEmpty", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Clears all condition sets to an empty collection. None status icons will be priorized. ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizer_ResetToEmpty_Description {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizer_ResetToEmpty_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die In Duty ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizerConditionSets_InDuty {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizerConditionSets_InDuty", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Status icons that should get priorized within duties. ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizerConditionSets_InDuty_Description {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizerConditionSets_InDuty_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die In Foray ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizerConditionSets_InForay {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizerConditionSets_InForay", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Status icons that should get priorized within foraies. ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizerConditionSets_InForay_Description {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizerConditionSets_InForay_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Overworld ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizerConditionSets_Overworld {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizerConditionSets_Overworld", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Status icons that should get priorized at overworld. ähnelt.
+        /// </summary>
+        public static string Loc_StatusIconPriorizerConditionSets_Overworld_Description {
+            get {
+                return ResourceManager.GetString("Loc_StatusIconPriorizerConditionSets_Overworld_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Busy ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_Busy {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_Busy", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Disconnecting ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_Disconnecting {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_Disconnecting", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Duty Finder ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_DutyFinder {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_DutyFinder", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Group Pose ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_GroupPose {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_GroupPose", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Idle ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_Idle {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_Idle", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die In Duty ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_InDuty {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_InDuty", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Mentor ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_Mentor {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_Mentor", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Mentor Crafting ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_MentorCrafting {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_MentorCrafting", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Mentor PvE ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_MentorPvE {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_MentorPvE", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Mentor PvP ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_MentorPvP {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_MentorPvP", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die New Adventurer ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_NewAdventurer {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_NewAdventurer", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Party Leader ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_PartyLeader {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_PartyLeader", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Party Member ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_PartyMember {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_PartyMember", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Returner ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_Returner {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_Returner", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Role Playing ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_RolePlaying {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_RolePlaying", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Viewing Cutscene ähnelt.
+        /// </summary>
+        public static string Loc_StatusIcons_ViewingCutscene {
+            get {
+                return ResourceManager.GetString("Loc_StatusIcons_ViewingCutscene", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die After ähnelt.
         /// </summary>
         public static string Loc_TagPosition_After {
             get {
@@ -1303,7 +1968,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Display the tag after the target element..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Display the tag after the target element. ähnelt.
         /// </summary>
         public static string Loc_TagPosition_After_Description {
             get {
@@ -1312,7 +1977,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Before.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Before ähnelt.
         /// </summary>
         public static string Loc_TagPosition_Before {
             get {
@@ -1321,7 +1986,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Display the tag before the target element..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Display the tag before the target element. ähnelt.
         /// </summary>
         public static string Loc_TagPosition_Before_Description {
             get {
@@ -1330,7 +1995,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Replace.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Replace ähnelt.
         /// </summary>
         public static string Loc_TagPosition_Replace {
             get {
@@ -1339,7 +2004,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Replace the target element with the tag..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Replace the target element with the tag. ähnelt.
         /// </summary>
         public static string Loc_TagPosition_Replace_Description {
             get {
@@ -1348,7 +2013,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Position in chat.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Position in chat ähnelt.
         /// </summary>
         public static string Loc_TagPositionInChat {
             get {
@@ -1357,7 +2022,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Where the tag will be positioned in chat..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Where the tag will be positioned in chat. ähnelt.
         /// </summary>
         public static string Loc_TagPositionInChat_Description {
             get {
@@ -1366,7 +2031,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Position in nameplates.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Position in nameplates ähnelt.
         /// </summary>
         public static string Loc_TagPositionInNameplates {
             get {
@@ -1375,7 +2040,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Where the tag will be positioned in nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die Where the tag will be positioned in nameplates. ähnelt.
         /// </summary>
         public static string Loc_TagPositionInNameplates_Description {
             get {
@@ -1384,7 +2049,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Target in nameplates.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Target in nameplates ähnelt.
         /// </summary>
         public static string Loc_TagTargetInNameplates {
             get {
@@ -1393,7 +2058,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The element that the tag should target in nameplates..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The element that the tag should target in nameplates. ähnelt.
         /// </summary>
         public static string Loc_TagTargetInNameplates_Description {
             get {
@@ -1402,7 +2067,44 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Text.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Target Chat Types ähnelt.
+        /// </summary>
+        public static string Loc_TargetChatTypes {
+            get {
+                return ResourceManager.GetString("Loc_TargetChatTypes", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Defines for which chat type the chat features of this tag should be enabled for. ähnelt.
+        /// </summary>
+        public static string Loc_TargetChatTypes_Description {
+            get {
+                return ResourceManager.GetString("Loc_TargetChatTypes_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Include undefined Chat Types ähnelt.
+        /// </summary>
+        public static string Loc_TargetChatTypesIncludeUndefined {
+            get {
+                return ResourceManager.GetString("Loc_TargetChatTypesIncludeUndefined", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die When enabled the Tag will apply to all Chat messages that has not a defined type.
+        ///This case can happen either if the Game updates and the Enumeration of all Chat Types gets invalid due to shifted values, or plugins creates a custom chat type for whatever reason. ähnelt.
+        /// </summary>
+        public static string Loc_TargetChatTypesIncludeUndefined_Description {
+            get {
+                return ResourceManager.GetString("Loc_TargetChatTypesIncludeUndefined_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Text ähnelt.
         /// </summary>
         public static string Loc_Text {
             get {
@@ -1411,7 +2113,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The text that will be displayed..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The text that will be displayed. ähnelt.
         /// </summary>
         public static string Loc_Text_Description {
             get {
@@ -1420,7 +2122,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Text properties.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Text properties ähnelt.
         /// </summary>
         public static string Loc_TextCategory {
             get {
@@ -1429,7 +2131,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Color.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Color ähnelt.
         /// </summary>
         public static string Loc_TextColor {
             get {
@@ -1438,7 +2140,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The color of the text..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The color of the text. ähnelt.
         /// </summary>
         public static string Loc_TextColor_Description {
             get {
@@ -1447,7 +2149,7 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to Glow color.
+        ///   Sucht eine lokalisierte Zeichenfolge, die Glow color ähnelt.
         /// </summary>
         public static string Loc_TextGlowColor {
             get {
@@ -1456,11 +2158,30 @@ namespace PlayerTags.Resources {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to The glow color of the text..
+        ///   Sucht eine lokalisierte Zeichenfolge, die The glow color of the text. ähnelt.
         /// </summary>
         public static string Loc_TextGlowColor_Description {
             get {
                 return ResourceManager.GetString("Loc_TextGlowColor_Description", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die Use priorized icons ähnelt.
+        /// </summary>
+        public static string Loc_UsePriorizedIcons {
+            get {
+                return ResourceManager.GetString("Loc_UsePriorizedIcons", resourceCulture);
+            }
+        }
+        
+        /// <summary>
+        ///   Sucht eine lokalisierte Zeichenfolge, die This option checked will force a set of status icons to be priorized over job icons, if you enabled job icons.
+        ///When disabled, only the disconnected status icon is priorized. ähnelt.
+        /// </summary>
+        public static string Loc_UsePriorizedIcons_Description {
+            get {
+                return ResourceManager.GetString("Loc_UsePriorizedIcons_Description", resourceCulture);
             }
         }
     }
